+++ packages/iml-query/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## Unreleased
+
+## v0.7.0 - 2026.02.27
+
+- enhance import statement parsing
+- add multi-file import resolution
+- add top-level APIs for simplified manipulation
+
 ## v0.6.0 - 2026.02.24
 
 - added